feat: ads for middle of article sections (#360)
Branch Commits:
- moved adsense out of helpers
- moved helper functions into utils file
- removed advertising text in the block
- display ad for article middle section

Jira Issue:
https://pethealthinc.atlassian.net/browse/PM-87

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After:
  - https://feat-pm-87-article-middle-ads--petplace--hlxsites.hlx.live/
-
https://feat-pm-87-article-middle-ads--petplace--hlxsites.hlx.live/?martech=off

---------

(cherry picked from commit f91acbbd3e33dff8caa079c0486d9e0e92344c00)
