+++ templates/article-page/article-page.js
@@ -10,7 +10,7 @@ import {
   getCategories,
   getCategory,
 } from '../../scripts/scripts.js';
-import { adsDefineSlot, adsDivCreator } from '../../scripts/utils/adsense.js';
+import { adsDefineSlot, adsDivCreator } from '../../scripts/adsense.js';
 
 export async function getCategoryByKey(key, value) {
   const categories = await getCategories();
@@ -214,11 +214,17 @@ export async function loadLazy(document) {
   breadcrumb.style.visibility = '';
 }
 
-// top, bottom, anchor
-// (middle + side later)
+// top, middle, bottom, anchor
+// (side later with refactor)
 export function loadDelayed() {
   adsDivCreator('article_top');
+  adsDivCreator('article_middle');
   adsDivCreator('article_bottom');
 
-  adsDefineSlot('article_top', 'article_bottom', 'article_anchor');
+  adsDefineSlot(
+    'article_top',
+    'article_middle',
+    'article_bottom',
+    'article_anchor',
+  );
 }
